send user data to api
