+++ Includes/functions.php
@@ -5,8 +5,6 @@
  * @package WPBP
  */
 
-use STOREGROWTH\SPSB\Admin\WPInsightsStoreGrowthSalesBooster;
-
 // If this file is called directly, abort.
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -130,40 +128,4 @@ if ( ! function_exists( 'sgsb_sanitize_svg_icon_fields' ) ) {
 
 		return wp_kses( $value, $icon_allowed_html );
 	}
-}
-
-
-if ( ! function_exists( 'storegrowth_sales_booster_start_plugin_tracking' ) ) {
-
-	/**
-	 * This function allows you to track usage of your plugin
-	 */
-	function storegrowth_sales_booster_start_plugin_tracking() {
-		$tracker = WPInsightsStoreGrowthSalesBooster::get_instance(
-			plugin_dir_path( __FILE__ ),
-			array(
-				'opt_in'       => true,
-				'goodbye_form' => true,
-				'item_id'      => 'cdabb75e4451684da3a3',
-			)
-		);
-
-		$tracker->set_notice_options(
-			array(
-				'notice'       => __( 'Never miss an important update!
- Opt in to get email notifications for security & feature updates, educational content, and occasional offers, and to share some basic WordPress environment info.', 'storegrowth-sales-booster' ),
-				'extra_notice' => __(
-					'We collect non-sensitive diagnostic data and plugin usage information. 
-        Your site URL, WordPress & PHP version, plugins & themes and email address to send you the 
-        discount coupon. This data lets us make sure this plugin always stays compatible with the most 
-        popular plugins and themes. No spam, I promise.',
-					'storegrowth-sales-booster'
-				),
-			)
-		);
-
-		$tracker->init();
-	}
-
-	storegrowth_sales_booster_start_plugin_tracking();
 }
